rename attachment handler modules
